Lite version of the application (WIP)
